test initial refactor of remote-lsp
